requirements.txt
python-dotenv==1.0.1
vkbottle==4.4.2